feat: login response 데이터 추가 및 수정
- nickname(닉네임), birthYear(유저 출생 년도) 추가
- gender(성별) -> MAN, WOMAN
- companyCategory(회사 카테고리)
 : MAJOR_COMPANY, MAJOR_COMPANY_SUBSIDIARY, MEDIUM_COMPANY , ...

+++ src/main/java/com/moa/company/infrastructure/CompanyRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface CompanyRepository extends JpaRepository<Company, Long> {
+public interface CompanyRepository extends JpaRepository<Company, Integer> {
 
 	Optional<Company> findByCompanyEmail(String companyEmail);
